Separate general Bitmap code from Outpost 2 specific bitmap code (sprites)
