+++ OP2Utility.vcxproj
@@ -122,6 +122,11 @@
     <ClInclude Include="src\Archives\CompressionType.h" />
     <ClInclude Include="src\Archives\WaveFile.h" />
     <ClInclude Include="include\OP2Utility.h" />
+    <ClInclude Include="src\Bitmaps\BmpCompression.h" />
+    <ClInclude Include="src\Bitmaps\BmpHeader.h" />
+    <ClInclude Include="src\Bitmaps\Color.h" />
+    <ClInclude Include="src\Bitmaps\ImageHeader.h" />
+    <ClInclude Include="src\Bitmaps\IndexedBmpWriter.h" />
     <ClInclude Include="src\Maps\CellType.h" />
     <ClInclude Include="src\Maps\MapData.h" />
     <ClInclude Include="src\Maps\MapHeader.h" />
@@ -144,6 +149,10 @@
     <ClCompile Include="src\Archives\ArchivePacker.cpp" />
     <ClCompile Include="src\Archives\ArchiveUnpacker.cpp" />
     <ClCompile Include="src\Archives\WaveFile.cpp" />
+    <ClCompile Include="src\Bitmaps\BmpHeader.cpp" />
+    <ClCompile Include="src\Bitmaps\Color.cpp" />
+    <ClCompile Include="src\Bitmaps\ImageHeader.cpp" />
+    <ClCompile Include="src\Bitmaps\IndexedBmpWriter.cpp" />
     <ClCompile Include="src\Maps\MapData.cpp" />
     <ClCompile Include="src\Maps\MapHeader.cpp" />
     <ClCompile Include="src\Maps\MapReader.cpp" />
@@ -157,10 +166,6 @@
     <ClCompile Include="src\Sprites\ArtReader.cpp" />
     <ClCompile Include="src\Sprites\ArtFile.cpp" />
     <ClCompile Include="src\Sprites\ArtWriter.cpp" />
-    <ClCompile Include="src\Sprites\BmpHeader.cpp" />
-    <ClCompile Include="src\Sprites\Color.cpp" />
-    <ClCompile Include="src\Sprites\IndexedBmpWriter.cpp" />
-    <ClCompile Include="src\Sprites\ImageHeader.cpp" />
     <ClCompile Include="src\Sprites\OP2BmpLoader.cpp" />
     <ClCompile Include="src\Sprites\PaletteHeader.cpp" />
     <ClCompile Include="src\Sprites\SectionHeader.cpp" />
@@ -171,11 +176,6 @@
     <ClCompile Include="src\Streams\MemoryReader.cpp" />
     <ClInclude Include="src\Sprites\Animation.h" />
     <ClInclude Include="src\Sprites\ArtFile.h" />
-    <ClInclude Include="src\Sprites\BmpCompression.h" />
-    <ClInclude Include="src\Sprites\BmpHeader.h" />
-    <ClInclude Include="src\Sprites\IndexedBmpWriter.h" />
-    <ClInclude Include="src\Sprites\Color.h" />
-    <ClInclude Include="src\Sprites\ImageHeader.h" />
     <ClInclude Include="src\Sprites\ImageMeta.h" />
     <ClInclude Include="src\Sprites\OP2BmpLoader.h" />
     <ClInclude Include="src\Sprites\PaletteHeader.h" />
