+++ src/Bitmaps/BmpCompression.h
