+++ src/Bitmaps/BmpHeader.cpp
