+++ src/Bitmaps/BmpHeader.h
