+++ src/Bitmaps/Color.cpp
