+++ src/Bitmaps/Color.h
