+++ src/Bitmaps/ImageHeader.cpp
