+++ src/Bitmaps/ImageHeader.h
