+++ src/Bitmaps/IndexedBmpWriter.cpp
