+++ src/Bitmaps/IndexedBmpWriter.h
