+++ test/OP2UtilityTest.vcxproj
@@ -42,15 +42,15 @@
   </PropertyGroup>
   <ItemGroup>
     <ClCompile Include="Archives\AdaptiveHuffmanTree.test.cpp" />
+    <ClCompile Include="Bitmaps\BmpHeader.test.cpp" />
+    <ClCompile Include="Bitmaps\Color.test.cpp" />
+    <ClCompile Include="Bitmaps\ImageHeader.test.cpp" />
+    <ClCompile Include="Bitmaps\IndexedBmpWriter.test.cpp" />
     <ClCompile Include="Maps\MapHeader.test.cpp" />
     <ClCompile Include="Maps\MapReader.test.cpp" />
     <ClCompile Include="Maps\MapWriter.test.cpp" />
     <ClCompile Include="Sprites\ArtReader.test.cpp" />
     <ClCompile Include="Sprites\ArtWriter.test.cpp" />
-    <ClCompile Include="Sprites\BmpHeader.test.cpp" />
-    <ClCompile Include="Sprites\Color.test.cpp" />
-    <ClCompile Include="Sprites\ImageHeader.test.cpp" />
-    <ClCompile Include="Sprites\IndexedBmpWriter.test.cpp" />
     <ClCompile Include="Streams\FileReader.test.cpp" />
     <ClCompile Include="Streams\FileSliceReader.test.cpp" />
     <ClCompile Include="Streams\MemoryStreamReader.test.cpp" />
